Do aggregation inline
