+++ discovery-provider/src/models/__init__.py
@@ -58,7 +58,6 @@ from .models import (
     UserListeningHistory,
     WalletChain,
 )
-from .persistent_store_entry import PersistentStoreEntry
 from .related_artist import RelatedArtist
 from .reward_manager import RewardManagerTransaction
 from .spl_token_transaction import SPLTokenTransaction
@@ -102,7 +101,6 @@ __all__ = [
     "IndexingCheckpoints",
     "ListenStreakChallenge",
     "Milestone",
-    "PersistentStoreEntry",
     "Play",
     "PlaysArchive",
     "Playlist",
